Refactor InputMessage and SendButton components to support customizable placeholders and disabled state; rename send-chat-thread index file for consistency
