+++ src/features/chats/chat-threads/index.tsx
@@ -2,17 +2,22 @@ import { ChatEvent } from '@/features/chats';
 
 import { Spinner } from '@/shared/components/spinner';
 
+import { SendChatThread } from '@/features/chats/chat-threads/components/send-chat-thread';
 import { useChatThreads } from './hooks';
 
 export const ChatThreads = () => {
   const { isLoadingThreads, threads } = useChatThreads();
 
-  if (isLoadingThreads) return <Spinner />;
-
-  if (!isLoadingThreads && !threads) return <div>No threads found</div>;
-
   return (
-    <div className="overflow-y-auto flex flex-col items-center p-2">
+    <div className="overflow-y-auto flex flex-col items-center p-2 gap-2">
+      <SendChatThread />
+
+      {isLoadingThreads && <Spinner />}
+
+      {!isLoadingThreads && !threads && (
+        <div className="w-full p-2 text-center">No threads found</div>
+      )}
+
       <div className="flex flex-col-reverse gap-2">
         {threads?.map((thread) => <ChatEvent key={thread.id} event={thread.id} />)}
       </div>
